deprecate index based foreach
