+++ Sources/RxComposableArchitecture/Reducer.swift
@@ -292,100 +292,6 @@ public struct Reducer<State, Action, Environment> {
         }
     }
 
-    /// A version of `pullback` that transforms a reducer that works on an element into one that works
-    /// on a collection of elements.
-    ///
-    ///     // Global domain that holds a collection of local domains:
-    ///     struct AppState { var todos: [Todo] }
-    ///     enum AppAction { case todo(index: Int, action: TodoAction) }
-    ///     struct AppEnvironment { var mainQueue: AnySchedulerOf<DispatchQueue> }
-    ///
-    ///     // A reducer that works on a local domain:
-    ///     let todoReducer = Reducer<Todo, TodoAction, TodoEnvironment> { ... }
-    ///
-    ///     // Pullback the local todo reducer so that it works on all of the app domain:
-    ///     let appReducer = Reducer<AppState, AppAction, AppEnvironment>.combine(
-    ///       todoReducer.forEach(
-    ///         state: \.todos,
-    ///         action: /AppAction.todo(index:action:),
-    ///         environment: { _ in TodoEnvironment() }
-    ///       ),
-    ///       Reducer { state, action, environment in
-    ///         ...
-    ///       }
-    ///     )
-    ///
-    /// Take care when combining `forEach` reducers into parent domains, as order matters. Always
-    /// combine `forEach` reducers _before_ parent reducers that can modify the collection.
-    ///
-    /// - Parameters:
-    ///   - toLocalState: A key path that can get/set an array of `State` elements inside.
-    ///     `GlobalState`.
-    ///   - toLocalAction: A case path that can extract/embed `(Int, Action)` from `GlobalAction`.
-    ///   - toLocalEnvironment: A function that transforms `GlobalEnvironment` into `Environment`.
-    ///   - breakpointOnNil: Raises `SIGTRAP` signal when an action is sent to the reducer but the
-    ///     index is out of bounds. This is generally considered a logic error, as a child reducer
-    ///     cannot process a child action for unavailable child state.
-    /// - Returns: A reducer that works on `GlobalState`, `GlobalAction`, `GlobalEnvironment`.
-    public func forEach<GlobalState, GlobalAction, GlobalEnvironment>(
-        state toLocalState: WritableKeyPath<GlobalState, [State]>,
-        action toLocalAction: CasePath<GlobalAction, (Int, Action)>,
-        environment toLocalEnvironment: @escaping (GlobalEnvironment) -> Environment,
-        breakpointOnNil: Bool = true,
-        _ file: StaticString = #file,
-        _ line: UInt = #line
-    ) -> Reducer<GlobalState, GlobalAction, GlobalEnvironment> {
-        .init { globalState, globalAction, globalEnvironment in
-            guard let (index, localAction) = toLocalAction.extract(from: globalAction) else {
-                return .none
-            }
-            if index >= globalState[keyPath: toLocalState].endIndex {
-                #if DEBUG
-                    if breakpointOnNil {
-                        fputs(
-                            """
-                            ---
-                            Warning: Reducer.forEach@\(file):\(line)
-
-                            "\(debugCaseOutput(localAction))" was received by a "forEach" reducer at index \
-                            \(index) when its state contained no element at this index. This is generally \
-                            considered an application logic error, and can happen for a few reasons:
-
-                            * This "forEach" reducer was combined with or run from another reducer that removed \
-                            the element at this index when it handled this action. To fix this make sure that \
-                            this "forEach" reducer is run before any other reducers that can move or remove \
-                            elements from state. This ensures that "forEach" reducers can handle their actions \
-                            for the element at the intended index.
-
-                            * An in-flight effect emitted this action while state contained no element at this \
-                            index. While it may be perfectly reasonable to ignore this action, you may want to \
-                            cancel the associated effect when moving or removing an element. If your "forEach" \
-                            reducer returns any long-living effects, you should use the identifier-based \
-                            "forEach" instead.
-
-                            * This action was sent to the store while its state contained no element at this \
-                            index. To fix this make sure that actions for this reducer can only be sent to a \
-                            view store when its state contains an element at this index. In SwiftUI \
-                            applications, use "ForEachStore".
-                            ---
-
-                            """,
-                            stderr
-                        )
-                        raise(SIGTRAP)
-                    }
-                #endif
-                return .none
-            }
-            return self.reducer(
-                &globalState[keyPath: toLocalState][index],
-                localAction,
-                toLocalEnvironment(globalEnvironment)
-            )
-            .map { toLocalAction.embed((index, $0)) }
-        }
-    }
-
     public func forEach<GlobalState, GlobalAction, GlobalEnvironment, Key>(
         state toLocalState: WritableKeyPath<GlobalState, [Key: State]>,
         action toLocalAction: CasePath<GlobalAction, (Key, Action)>,
@@ -587,38 +493,5 @@ extension Reducer where Environment == Void {
         _ action: Action
     ) -> Effect<Action> {
         callAsFunction(&state, action, ())
-    }
-}
-
-extension Reducer where State: HashDiffable {
-    /**
-     Pullback reducer for reducer which store is part of `Differentiable` array.
-     */
-    public func forEach<Identifier, GlobalState, GlobalAction, GlobalEnvironment>(
-        state toLocalState: WritableKeyPath<GlobalState, [State]>,
-        action toLocalAction: CasePath<GlobalAction, (Identifier, Action)>,
-        environment toLocalEnvironment: @escaping (GlobalEnvironment) -> Environment
-    ) -> Reducer<GlobalState, GlobalAction, GlobalEnvironment>
-        where Identifier == State.IdentifierType {
-        .init { globalState, globalAction, globalEnvironment in
-            guard let (identifier, localAction) = toLocalAction.extract(from: globalAction) else {
-                return .none
-            }
-
-            // search index of identifier
-            guard let index = globalState[keyPath: toLocalState].firstIndex(where: { $0.id == identifier })
-            else {
-                assertionFailure("\(identifier) is not exist on Global State")
-                return .none
-            }
-
-            // return redux
-            return self.reducer(
-                &globalState[keyPath: toLocalState][index],
-                localAction,
-                toLocalEnvironment(globalEnvironment)
-            )
-            .map { toLocalAction.embed((identifier, $0)) }
-        }
     }
 }
